scaling is actually working now

+++ plotter.cpp
@@ -15,7 +15,7 @@ Plotter::~Plotter()
 void Plotter::plotData(QRect &canvas, QPainter &painter, float *pData, float intervalLength)
 {
     float step = intervalLength / RenderScreen::m_dataSize;
-    int x_scale = canvas.width() / intervalLength;
+    float x_scale = canvas.width() / intervalLength;
 
     float max = pData[0];
     for (int i = 0; i < RenderScreen::m_dataSize; ++i)
@@ -23,7 +23,7 @@ void Plotter::plotData(QRect &canvas, QPainter &painter, float *pData, float int
         if (pData[i] > max) max = pData[i];
     }
 
-    int y_scale = canvas.height() / (4 * max);
+    float y_scale = canvas.height() / (4 * max);
 
     QPoint previousPixel = calculatePoint(0, step, canvas, pData[0], x_scale, y_scale);
     for (int i = 0; i < RenderScreen::m_dataSize; ++i)
